session lumen + validacao front

+++ vuejs/src/components/shared/Navbar.vue
@@ -1,7 +1,7 @@
 <template>
   <div>
-    <b-navbar toggleable="lg" type="dark" variant="dark">
-      <b-navbar-brand href="#">
+    <b-navbar toggleable="lg" type="dark" variant="dark" style="--bs-bg-opacity: .5;">
+      <b-navbar-brand href="/">
         <img height="30" src="../../assets/img/logo.png"/>
         Spotify</b-navbar-brand>
 
@@ -9,15 +9,19 @@
 
       <b-collapse id="nav-collapse" is-nav>
         <b-navbar-nav>
-          <b-nav-item href="#">Bandas</b-nav-item>
+          <b-nav-item class="nav_item" href="#">Bandas</b-nav-item>
           <b-nav-item href="#" disabled>|</b-nav-item>
           <b-nav-item href="#">Gênero</b-nav-item>
+          <b-navbar-nav v-if="isLogged">
+            <b-nav-item href="#" disabled>|</b-nav-item>
+            <b-nav-item href="#">Músicas favoritas</b-nav-item>
+          </b-navbar-nav>
         </b-navbar-nav>
 
         <!-- Right aligned nav items -->
         <b-navbar-nav class="ml-auto">
           <b-nav-form>
-            <b-form-input size="sm" class="mr-lg-5" placeholder="Música, álbum ou banda"></b-form-input>
+            <b-form-input size="sm" class="mr-lg-5 search-sptf" placeholder="Música, álbum ou banda"></b-form-input>
           </b-nav-form>
 
           <b-navbar-nav right v-if="!isLogged">
@@ -31,10 +35,14 @@
           <b-nav-item-dropdown right v-else>
             <!-- Using 'button-content' slot -->
             <template #button-content>
-              {{ userName }}
+              Perfil
             </template>
-            <b-dropdown-item href="#">Profile</b-dropdown-item>
-            <b-dropdown-item href="#" @click="logout()">Sign Out</b-dropdown-item>
+            <b-dropdown-item href="#" v-b-modal.modal-3>Visualizar perfil</b-dropdown-item>
+            <b-dropdown-item href="#" @click="logout()">Logout</b-dropdown-item>
+            <div v-if="accessLevel > 0">
+              <hr/>
+              <b-dropdown-item href="/admin">Painel administrativo</b-dropdown-item>
+            </div>
           </b-nav-item-dropdown>
 
         </b-navbar-nav>
@@ -47,6 +55,13 @@
 
     <b-modal id="modal-2" hide-footer title="Inscreva-se">
       <FormLogin/>
+    </b-modal>
+
+    <b-modal id="modal-3" hide-footer :title="`Perfil de ${userName}`">
+      <b-card>
+        <p><strong>Nome:</strong> {{ userName }}</p>
+        <p><strong>Email:</strong> {{ userEmail }}</p>
+      </b-card>
     </b-modal>
   </div>
 </template>
@@ -65,6 +80,7 @@ export default {
       isLogged: false,
       userName: '',
       userEmail: '',
+      accessLevel: 0,
     }
   },
   created() {
@@ -72,15 +88,21 @@ export default {
   },
   methods: {
     validateLogin(){
-      if(sessionStorage.getItem('user_email')){
-        this.isLogged = true
+      if(sessionStorage.getItem('is_logged')){
+        this.isLogged = sessionStorage.getItem('is_logged')
         this.userEmail = sessionStorage.getItem('user_email')
         this.userName = sessionStorage.getItem('user_name')
+        this.accessLevel = sessionStorage.getItem('user_access_level')
       }
     },
     logout(){
-      sessionStorage.clear()
-      location.reload()
+      this.$axios.post('/auth/logout')
+      .then(res => {
+        if(res.status == 200){
+          sessionStorage.clear()
+          location.reload()
+        }
+      })
     }
   }
 }
@@ -89,5 +111,22 @@ export default {
 <style scoped>
 .nv {
   margin-left: 1% !important;
+}
+
+.bg-dark {
+  background-color: #1c1c1c73 !important;
+}
+
+.search-sptf{
+  border-radius: 80px;
+  opacity: 50%;
+  background-color: rgba(58, 52, 52, 0.66);
+  -webkit-border-radius: 80px;
+  -moz-border-radius: 80px;
+  -ms-border-radius: 80px;
+  -o-border-radius: 80px;
+}
+.search-sptf::placeholder {
+  color: #ffffff;
 }
 </style>
